+++ testutils/udp_transport.go
@@ -19,7 +19,7 @@ import (
 	"net"
 	"sync/atomic"
 
-	thrift "github.com/uber/jaeger-client-go/internal/thrift_0_10"
+	"github.com/uber/jaeger-client-go/thrift"
 )
 
 const (
