+++ thrift/binary_protocol.go
