+++ thrift/compact_protocol.go
