+++ thrift/memory_buffer.go
