+++ thrift/processor.go
