+++ thrift/rich_transport.go
