+++ thrift/simple_json_protocol.go
