+++ thrift/type.go
